PersonaWay API/requirements.txt
Flask
mysql-connector-python
google-cloud-storage
tensorflow
numpy
h5py